refactor #step2 jdbcTemplate 리팩토링
