+++ src/main/java/core/jdbc/JdbcOperations.java
@@ -1,13 +1,12 @@
 package core.jdbc;
 
 import java.util.List;
-import java.util.Optional;
 
 public interface JdbcOperations {
 
     void execute(String sql, Object... parameters);
 
-    <T> List<T> queryForList(String sql, ResultSetExtractor<T> requestType);
+    <T> List<T> queryForList(String sql, ResultSetExtractor<T> requestType, Object... parameters);
 
     <T> T queryForObject(String sql, ResultSetExtractor<T> resultSetExtractor, Object... parameters);
 }
