+++ src/main/java/core/jdbc/JdbcTemplate.java
@@ -1,8 +1,8 @@
 package core.jdbc;
 
 
-import next.model.User;
-
+import java.sql.Connection;
+import java.sql.PreparedStatement;
 import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.util.ArrayList;
@@ -10,44 +10,50 @@ import java.util.List;
 
 public class JdbcTemplate implements JdbcOperations {
 
-    @Override
-    public void execute(String sql, Object... parameters) {
-        ExecuteTemplate template = new ExecuteTemplate() {
-            @Override
-            <R> R getResultMapper(ResultSetExtractor extractor, ResultSet rs) {
-                return null;
-            }
-        };
-        template.execute(sql, parameters);
-    }
+    private static final int START_PARAMETER_NUMBER = 1;
+    public static final int ONE_OBJECT = 0;
 
     @Override
-    public <T> List<T> queryForList(String sql, ResultSetExtractor<T> resultSetExtractor) {
-        ExecuteTemplate template = new ExecuteTemplate() {
-            @Override
-            List<User> getResultMapper(ResultSetExtractor extractor, ResultSet rs) throws SQLException {
-                List<User> results = new ArrayList<>();
-                while (rs.next()) {
-                    results.add((User) resultSetExtractor.extractData(rs));
-                }
-                return results;
-            }
-        };
-        return template.execute(sql, resultSetExtractor);
+    public void execute(String sql, Object... parameters) {
+        execute(PreparedStatement::executeUpdate, sql, parameters);
     }
 
     @Override
     public <T> T queryForObject(String sql, ResultSetExtractor<T> resultSetExtractor, Object... parameters) {
-        ExecuteTemplate template = new ExecuteTemplate() {
-            @Override
-            User getResultMapper(ResultSetExtractor extractor, ResultSet rs) throws SQLException {
-                List<User> results = new ArrayList<>();
-                while (rs.next()) {
-                    results.add((User) resultSetExtractor.extractData(rs));
-                }
-                return results.get(0);
+        return queryForList(sql, resultSetExtractor, parameters).get(ONE_OBJECT);
+    }
+
+    @Override
+    public <T> List<T> queryForList(String sql, ResultSetExtractor<T> resultSetExtractor, Object... parameters) {
+        return query(sql, resultSetExtractor, parameters);
+    }
+
+    <R> R execute(StatementCallback<R> callback, String sql, Object... parameters) {
+        try (Connection connection = ConnectionManager.getConnection();
+             PreparedStatement ps = connection.prepareStatement(sql)) {
+            setValues(ps, parameters);
+
+            return callback.perform(ps);
+        } catch (SQLException e) {
+            throw new JdbcExecuteException(e);
+        }
+    }
+
+    <R> List<R> query(String sql, ResultSetExtractor<R> resultSetExtractor, Object... parameters) {
+        return execute(ps -> {
+            ResultSet rs = ps.executeQuery();
+            List<R> results = new ArrayList<>();
+            while (rs.next()) {
+                results.add(resultSetExtractor.extract(rs));
             }
-        };
-        return template.execute(sql, resultSetExtractor, parameters);
+            return results;
+        }, sql, parameters);
+    }
+
+    private void setValues(PreparedStatement ps, Object[] parameters) throws SQLException {
+        int i = START_PARAMETER_NUMBER;
+        for (Object obj : parameters) {
+            ps.setString(i++, obj.toString());
+        }
     }
 }
