--- src/main/java/core/jdbc/PrepareStatementSetter.java
@@ -1,9 +1,0 @@
-package core.jdbc;
-
-import java.sql.PreparedStatement;
-import java.sql.SQLException;
-
-public interface PrepareStatementSetter {
-
-    void setValues(PreparedStatement ps) throws SQLException;
-}
