--- src/main/java/core/jdbc/PreparedStatementCreator.java
@@ -1,11 +1,0 @@
-package core.jdbc;
-
-import java.sql.Connection;
-import java.sql.PreparedStatement;
-import java.sql.SQLException;
-
-@FunctionalInterface
-public interface PreparedStatementCreator {
-
-    PreparedStatement create(Connection con) throws SQLException;
-}
