+++ src/main/java/core/jdbc/ResultSetExtractor.java
@@ -4,7 +4,7 @@ import java.sql.ResultSet;
 import java.sql.SQLException;
 
 @FunctionalInterface
-public interface ResultSetExtractor<T> {
+public interface ResultSetExtractor<R> {
 
-    T extractData(ResultSet rs) throws SQLException;
+    R extract(ResultSet rs) throws SQLException;
 }
